flask-app/src/requirements.txt
Flask
mysql-connector-python
pytest
flask_jwt_extended
bcrypt
PyJWT
coverage
asyncssh